@@ -1,12 +1,13 @@
 package cmd
 
 import (
-	"blocky/api"
 	"encoding/json"
 	"net/http"
 	"net/http/httptest"
 	"net/url"
 	"strconv"
+
+	"github.com/0xERR0R/blocky/api"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
