@@ -1,9 +1,9 @@
 package cmd
 
 import (
-	. "blocky/log"
 	"testing"
 
+	. "github.com/0xERR0R/blocky/log"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
