@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"blocky/log"
+	"github.com/0xERR0R/blocky/log"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
