@@ -1,3 +1,4 @@
+//go:generate go-enum -f=$GOFILE --marshal --names
 package config
 
 import (
@@ -9,40 +10,45 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	"time"
 
-	"blocky/log"
-
+	"github.com/0xERR0R/blocky/log"
 	"gopkg.in/yaml.v2"
 )
 
+// NetProtocol resolver protocol ENUM(
+// udp // Deprecated: use tcp+udp instead
+// tcp // Deprecated: use tcp+udp instead
+// tcp+udp // TCP and UDP protocols
+// tcp-tls // TCP-TLS protocol
+// https // HTTPS protocol
+// )
+type NetProtocol uint16
+
+// QueryLogType type of the query log ENUM(
+// none // use logger as fallback
+// mysql // MySQL or MariaDB database
+// csv // CSV file per day
+// csv-client // CSV file per day and client
+// )
+type QueryLogType int16
+
+type Duration time.Duration
+
 const (
 	validUpstream = `(?P<Host>(?:\[[^\]]+\])|[^\s/:]+):?(?P<Port>[^\s/:]*)?(?P<Path>/[^\s]*)?`
-	// NetUDP UDP protocol (deprecated)
-	NetUDP = "udp"
-
-	// NetTCP TCP protocol (deprecated)
-	NetTCP = "tcp"
-
-	// NetTCPUDP TCP and UDP protocols
-	NetTCPUDP = "tcp+udp"
-
-	// NetTCPTLS TCP-TLS protocol
-	NetTCPTLS = "tcp-tls"
-
-	// NetHTTPS HTTPS protocol
-	NetHTTPS = "https"
 )
 
 // nolint:gochecknoglobals
-var netDefaultPort = map[string]uint16{
-	NetTCPUDP: 53,
-	NetTCPTLS: 853,
-	NetHTTPS:  443,
+var netDefaultPort = map[NetProtocol]uint16{
+	NetProtocolTcpUdp: 53,
+	NetProtocolTcpTls: 853,
+	NetProtocolHttps:  443,
 }
 
 // Upstream is the definition of external DNS server
 type Upstream struct {
-	Net  string
+	Net  NetProtocol
 	Host string
 	Port uint16
 	Path string
@@ -123,13 +129,36 @@ func (c *CustomDNSMapping) UnmarshalYAML(unmarshal func(interface{}) error) erro
 	return nil
 }
 
+// UnmarshalYAML creates Duration from YAML. If no unit is used, uses minutes
+func (c *Duration) UnmarshalYAML(unmarshal func(interface{}) error) error {
+	var input string
+	if err := unmarshal(&input); err != nil {
+		return err
+	}
+
+	if minutes, err := strconv.Atoi(input); err == nil {
+		// duration is defined as number without unit
+		// use minutes to ensure back compatibility
+		*c = Duration(time.Duration(minutes) * time.Minute)
+		return nil
+	}
+
+	duration, err := time.ParseDuration(input)
+	if err == nil {
+		*c = Duration(duration)
+		return nil
+	}
+
+	return err
+}
+
 // ParseUpstream creates new Upstream from passed string in format [net]:host[:port][/path]
 func ParseUpstream(upstream string) (result Upstream, err error) {
 	if strings.TrimSpace(upstream) == "" {
 		return Upstream{}, nil
 	}
 
-	var n string
+	var n NetProtocol
 
 	n, upstream = extractNet(upstream)
 
@@ -164,59 +193,64 @@ func ParseUpstream(upstream string) (result Upstream, err error) {
 	return Upstream{Net: n, Host: host, Port: port, Path: path}, nil
 }
 
-func extractNet(upstream string) (string, string) {
-	if strings.HasPrefix(upstream, NetTCP+":") {
+func extractNet(upstream string) (NetProtocol, string) {
+	if strings.HasPrefix(upstream, NetProtocolTcp.String()+":") {
 		log.Log().Warnf("net prefix tcp is deprecated, using tcp+udp as default fallback")
 
-		return NetTCPUDP, strings.Replace(upstream, NetTCP+":", "", 1)
+		return NetProtocolTcpUdp, strings.Replace(upstream, NetProtocolTcp.String()+":", "", 1)
 	}
 
-	if strings.HasPrefix(upstream, NetUDP+":") {
+	if strings.HasPrefix(upstream, NetProtocolUdp.String()+":") {
 		log.Log().Warnf("net prefix udp is deprecated, using tcp+udp as default fallback")
-		return NetTCPUDP, strings.Replace(upstream, NetUDP+":", "", 1)
+		return NetProtocolTcpUdp, strings.Replace(upstream, NetProtocolUdp.String()+":", "", 1)
 	}
 
-	if strings.HasPrefix(upstream, NetTCPUDP+":") {
-		return NetTCPUDP, strings.Replace(upstream, NetTCPUDP+":", "", 1)
+	if strings.HasPrefix(upstream, NetProtocolTcpUdp.String()+":") {
+		return NetProtocolTcpUdp, strings.Replace(upstream, NetProtocolTcpUdp.String()+":", "", 1)
 	}
 
-	if strings.HasPrefix(upstream, NetTCPTLS+":") {
-		return NetTCPTLS, strings.Replace(upstream, NetTCPTLS+":", "", 1)
+	if strings.HasPrefix(upstream, NetProtocolTcpTls.String()+":") {
+		return NetProtocolTcpTls, strings.Replace(upstream, NetProtocolTcpTls.String()+":", "", 1)
 	}
 
-	if strings.HasPrefix(upstream, NetHTTPS+":") {
-		return NetHTTPS, strings.Replace(upstream, NetHTTPS+":", "", 1)
+	if strings.HasPrefix(upstream, NetProtocolHttps.String()+":") {
+		return NetProtocolHttps, strings.Replace(upstream, NetProtocolHttps.String()+":", "", 1)
 	}
 
-	return NetTCPUDP, upstream
+	return NetProtocolTcpUdp, upstream
 }
 
 const (
-	cfgDefaultPort           = "53"
-	cfgDefaultPrometheusPath = "/metrics"
+	cfgDefaultPort            = "53"
+	cfgDefaultPrometheusPath  = "/metrics"
+	cfgDefaultUpstreamTimeout = Duration(2 * time.Second)
+	cfgDefaultRefreshPeriod   = Duration(4 * time.Hour)
+	cfgDefaultDownloadTimeout = Duration(60 * time.Second)
 )
 
 // Config main configuration
 // nolint:maligned
 type Config struct {
-	Upstream     UpstreamConfig            `yaml:"upstream"`
-	CustomDNS    CustomDNSConfig           `yaml:"customDNS"`
-	Conditional  ConditionalUpstreamConfig `yaml:"conditional"`
-	Blocking     BlockingConfig            `yaml:"blocking"`
-	ClientLookup ClientLookupConfig        `yaml:"clientLookup"`
-	Caching      CachingConfig             `yaml:"caching"`
-	QueryLog     QueryLogConfig            `yaml:"queryLog"`
-	Prometheus   PrometheusConfig          `yaml:"prometheus"`
-	LogLevel     string                    `yaml:"logLevel"`
-	LogFormat    string                    `yaml:"logFormat"`
-	LogTimestamp bool                      `yaml:"logTimestamp"`
-	Port         string                    `yaml:"port"`
-	HTTPPort     string                    `yaml:"httpPort"`
-	HTTPSPort    string                    `yaml:"httpsPort"`
-	DisableIPv6  bool                      `yaml:"disableIPv6"`
-	CertFile     string                    `yaml:"httpsCertFile"`
-	KeyFile      string                    `yaml:"httpsKeyFile"`
-	BootstrapDNS Upstream                  `yaml:"bootstrapDns"`
+	Upstream        UpstreamConfig            `yaml:"upstream"`
+	UpstreamTimeout Duration                  `yaml:"upstreamTimeout"`
+	CustomDNS       CustomDNSConfig           `yaml:"customDNS"`
+	Conditional     ConditionalUpstreamConfig `yaml:"conditional"`
+	Blocking        BlockingConfig            `yaml:"blocking"`
+	ClientLookup    ClientLookupConfig        `yaml:"clientLookup"`
+	Caching         CachingConfig             `yaml:"caching"`
+	QueryLog        QueryLogConfig            `yaml:"queryLog"`
+	Prometheus      PrometheusConfig          `yaml:"prometheus"`
+	LogLevel        log.Level                 `yaml:"logLevel"`
+	LogFormat       log.FormatType            `yaml:"logFormat"`
+	LogPrivacy      bool                      `yaml:"logPrivacy"`
+	LogTimestamp    bool                      `yaml:"logTimestamp"`
+	Port            string                    `yaml:"port"`
+	HTTPPort        string                    `yaml:"httpPort"`
+	HTTPSPort       string                    `yaml:"httpsPort"`
+	DisableIPv6     bool                      `yaml:"disableIPv6"`
+	CertFile        string                    `yaml:"httpsCertFile"`
+	KeyFile         string                    `yaml:"httpsKeyFile"`
+	BootstrapDNS    Upstream                  `yaml:"bootstrapDns"`
 }
 
 // PrometheusConfig contains the config values for prometheus
@@ -257,7 +291,9 @@ type BlockingConfig struct {
 	WhiteLists        map[string][]string `yaml:"whiteLists"`
 	ClientGroupsBlock map[string][]string `yaml:"clientGroupsBlock"`
 	BlockType         string              `yaml:"blockType"`
-	RefreshPeriod     int                 `yaml:"refreshPeriod"`
+	BlockTTL          Duration            `yaml:"blockTTL"`
+	DownloadTimeout   Duration            `yaml:"downloadTimeout"`
+	RefreshPeriod     Duration            `yaml:"refreshPeriod"`
 }
 
 // ClientLookupConfig configuration for the client lookup
@@ -269,24 +305,31 @@ type ClientLookupConfig struct {
 
 // CachingConfig configuration for domain caching
 type CachingConfig struct {
-	MinCachingTime        int  `yaml:"minTime"`
-	MaxCachingTime        int  `yaml:"maxTime"`
-	MaxItemsCount         int  `yaml:"maxItemsCount"`
-	Prefetching           bool `yaml:"prefetching"`
-	PrefetchExpires       int  `yaml:"prefetchExpires"`
-	PrefetchThreshold     int  `yaml:"prefetchThreshold"`
-	PrefetchMaxItemsCount int  `yaml:"prefetchMaxItemsCount"`
+	MinCachingTime        Duration `yaml:"minTime"`
+	MaxCachingTime        Duration `yaml:"maxTime"`
+	MaxItemsCount         int      `yaml:"maxItemsCount"`
+	Prefetching           bool     `yaml:"prefetching"`
+	PrefetchExpires       Duration `yaml:"prefetchExpires"`
+	PrefetchThreshold     int      `yaml:"prefetchThreshold"`
+	PrefetchMaxItemsCount int      `yaml:"prefetchMaxItemsCount"`
 }
 
 // QueryLogConfig configuration for the query logging
 type QueryLogConfig struct {
-	Dir              string `yaml:"dir"`
-	PerClient        bool   `yaml:"perClient"`
-	LogRetentionDays uint64 `yaml:"logRetentionDays"`
+	// Deprecated
+	Dir string `yaml:"dir"`
+	// Deprecated
+	PerClient        bool         `yaml:"perClient"`
+	Target           string       `yaml:"target"`
+	Type             QueryLogType `yaml:"type"`
+	LogRetentionDays uint64       `yaml:"logRetentionDays"`
 }
 
-// NewConfig creates new config from YAML file
-func NewConfig(path string, mandatory bool) Config {
+// nolint:gochecknoglobals
+var config = &Config{}
+
+// LoadConfig creates new config from YAML file
+func LoadConfig(path string, mandatory bool) {
 	cfg := Config{}
 	setDefaultValues(&cfg)
 
@@ -296,28 +339,55 @@ func NewConfig(path string, mandatory bool) Config {
 		if errors.Is(err, os.ErrNotExist) && !mandatory {
 			// config file does not exist
 			// return config with default values
-			return cfg
+			config = &cfg
+			return
 		}
 
 		log.Log().Fatal("Can't read config file: ", err)
 	}
 
-	err = yaml.UnmarshalStrict(data, &cfg)
+	unmarshalConfig(data, cfg)
+}
+
+func unmarshalConfig(data []byte, cfg Config) {
+	err := yaml.UnmarshalStrict(data, &cfg)
 	if err != nil {
 		log.Log().Fatal("wrong file structure: ", err)
 	}
 
-	if cfg.LogFormat != log.CfgLogFormatText && cfg.LogFormat != log.CfgLogFormatJSON {
-		log.Log().Fatal("LogFormat should be 'text' or 'json'")
-	}
+	validateConfig(&cfg)
 
-	return cfg
+	config = &cfg
+}
+
+func validateConfig(cfg *Config) {
+	if cfg.QueryLog.Dir != "" {
+		log.Log().Warnf("queryLog.Dir is deprecated, use 'queryLog.target' instead")
+
+		if cfg.QueryLog.Target == "" {
+			cfg.QueryLog.Target = cfg.QueryLog.Dir
+		}
+
+		if cfg.QueryLog.Type == QueryLogTypeNone {
+			if cfg.QueryLog.PerClient {
+				cfg.QueryLog.Type = QueryLogTypeCsvClient
+			} else {
+				cfg.QueryLog.Type = QueryLogTypeCsv
+			}
+		}
+	}
+}
+
+// GetConfig returns the current config
+func GetConfig() *Config {
+	return config
 }
 
 func setDefaultValues(cfg *Config) {
 	cfg.Port = cfgDefaultPort
-	cfg.LogLevel = "info"
-	cfg.LogFormat = log.CfgLogFormatText
 	cfg.LogTimestamp = true
 	cfg.Prometheus.Path = cfgDefaultPrometheusPath
+	cfg.UpstreamTimeout = cfgDefaultUpstreamTimeout
+	cfg.Blocking.RefreshPeriod = cfgDefaultRefreshPeriod
+	cfg.Blocking.DownloadTimeout = cfgDefaultDownloadTimeout
 }
