@@ -1,11 +1,14 @@
 package config
 
 import (
-	. "blocky/log"
 	"io/ioutil"
 	"net"
 	"os"
+	"time"
 
+	"github.com/0xERR0R/blocky/helpertest"
+
+	. "github.com/0xERR0R/blocky/log"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
@@ -18,30 +21,35 @@ var _ = Describe("Config", func() {
 				err := os.Chdir("../testdata")
 				Expect(err).Should(Succeed())
 
-				cfg := NewConfig("config.yml", true)
+				LoadConfig("config.yml", true)
 
-				Expect(cfg.Port).Should(Equal("55555"))
-				Expect(cfg.Upstream.ExternalResolvers["default"]).Should(HaveLen(3))
-				Expect(cfg.Upstream.ExternalResolvers["default"][0].Host).Should(Equal("8.8.8.8"))
-				Expect(cfg.Upstream.ExternalResolvers["default"][1].Host).Should(Equal("8.8.4.4"))
-				Expect(cfg.Upstream.ExternalResolvers["default"][2].Host).Should(Equal("1.1.1.1"))
-				Expect(cfg.CustomDNS.Mapping.HostIPs).Should(HaveLen(2))
-				Expect(cfg.CustomDNS.Mapping.HostIPs["my.duckdns.org"][0]).Should(Equal(net.ParseIP("192.168.178.3")))
-				Expect(cfg.CustomDNS.Mapping.HostIPs["multiple.ips"][0]).Should(Equal(net.ParseIP("192.168.178.3")))
-				Expect(cfg.CustomDNS.Mapping.HostIPs["multiple.ips"][1]).Should(Equal(net.ParseIP("192.168.178.4")))
-				Expect(cfg.CustomDNS.Mapping.HostIPs["multiple.ips"][2]).Should(Equal(
+				Expect(config.Port).Should(Equal("55555"))
+				Expect(config.Upstream.ExternalResolvers["default"]).Should(HaveLen(3))
+				Expect(config.Upstream.ExternalResolvers["default"][0].Host).Should(Equal("8.8.8.8"))
+				Expect(config.Upstream.ExternalResolvers["default"][1].Host).Should(Equal("8.8.4.4"))
+				Expect(config.Upstream.ExternalResolvers["default"][2].Host).Should(Equal("1.1.1.1"))
+				Expect(config.CustomDNS.Mapping.HostIPs).Should(HaveLen(2))
+				Expect(config.CustomDNS.Mapping.HostIPs["my.duckdns.org"][0]).Should(Equal(net.ParseIP("192.168.178.3")))
+				Expect(config.CustomDNS.Mapping.HostIPs["multiple.ips"][0]).Should(Equal(net.ParseIP("192.168.178.3")))
+				Expect(config.CustomDNS.Mapping.HostIPs["multiple.ips"][1]).Should(Equal(net.ParseIP("192.168.178.4")))
+				Expect(config.CustomDNS.Mapping.HostIPs["multiple.ips"][2]).Should(Equal(
 					net.ParseIP("2001:0db8:85a3:08d3:1319:8a2e:0370:7344")))
-				Expect(cfg.Conditional.Mapping.Upstreams).Should(HaveLen(2))
-				Expect(cfg.Conditional.Mapping.Upstreams["fritz.box"]).Should(HaveLen(1))
-				Expect(cfg.Conditional.Mapping.Upstreams["multiple.resolvers"]).Should(HaveLen(2))
-				Expect(cfg.ClientLookup.Upstream.Host).Should(Equal("192.168.178.1"))
-				Expect(cfg.ClientLookup.SingleNameOrder).Should(Equal([]uint{2, 1}))
-				Expect(cfg.Blocking.BlackLists).Should(HaveLen(2))
-				Expect(cfg.Blocking.WhiteLists).Should(HaveLen(1))
-				Expect(cfg.Blocking.ClientGroupsBlock).Should(HaveLen(2))
+				Expect(config.Conditional.Mapping.Upstreams).Should(HaveLen(2))
+				Expect(config.Conditional.Mapping.Upstreams["fritz.box"]).Should(HaveLen(1))
+				Expect(config.Conditional.Mapping.Upstreams["multiple.resolvers"]).Should(HaveLen(2))
+				Expect(config.ClientLookup.Upstream.Host).Should(Equal("192.168.178.1"))
+				Expect(config.ClientLookup.SingleNameOrder).Should(Equal([]uint{2, 1}))
+				Expect(config.Blocking.BlackLists).Should(HaveLen(2))
+				Expect(config.Blocking.WhiteLists).Should(HaveLen(1))
+				Expect(config.Blocking.ClientGroupsBlock).Should(HaveLen(2))
+				Expect(config.Blocking.BlockTTL).Should(Equal(Duration(time.Minute)))
+				Expect(config.Blocking.RefreshPeriod).Should(Equal(Duration(2 * time.Hour)))
 
-				Expect(cfg.Caching.MaxCachingTime).Should(Equal(0))
-				Expect(cfg.Caching.MinCachingTime).Should(Equal(0))
+				Expect(config.Caching.MaxCachingTime).Should(Equal(Duration(0)))
+				Expect(config.Caching.MinCachingTime).Should(Equal(Duration(0)))
+
+				Expect(GetConfig()).Should(Not(BeNil()))
+
 			})
 		})
 		When("config file is malformed", func() {
@@ -55,16 +63,101 @@ var _ = Describe("Config", func() {
 				err = ioutil.WriteFile("config.yml", []byte("malformed_config"), 0600)
 				Expect(err).Should(Succeed())
 
-				defer func() { Log().ExitFunc = nil }()
-
-				var fatal bool
-
-				Log().ExitFunc = func(int) { fatal = true }
-
-				_ = NewConfig("config.yml", true)
-				Expect(fatal).Should(BeTrue())
+				helpertest.ShouldLogFatal(func() {
+					LoadConfig("config.yml", true)
+				})
 			})
 		})
+		When("duration is in wrong format", func() {
+			It("should log with fatal and exit", func() {
+				cfg := Config{}
+				data :=
+					`blocking:
+  refreshPeriod: wrongduration`
+				helpertest.ShouldLogFatal(func() {
+					unmarshalConfig([]byte(data), cfg)
+				})
+			})
+		})
+		When("CustomDNS hast wrong IP defined", func() {
+			It("should log with fatal and exit", func() {
+				cfg := Config{}
+				data :=
+					`customDNS:
+  mapping:
+    someDomain: 192.168.178.WRONG`
+				helpertest.ShouldLogFatal(func() {
+					unmarshalConfig([]byte(data), cfg)
+				})
+			})
+		})
+		When("Conditional mapping hast wrong defined upstreams", func() {
+			It("should log with fatal and exit", func() {
+				cfg := Config{}
+				data :=
+					`conditional:
+  mapping:
+    multiple.resolvers: udp:192.168.178.1,wongprotocol:4.4.4.4:53`
+				helpertest.ShouldLogFatal(func() {
+					unmarshalConfig([]byte(data), cfg)
+				})
+			})
+		})
+		When("Wrong upstreams are defined", func() {
+			It("should log with fatal and exit", func() {
+				cfg := Config{}
+				data :=
+					`upstream:
+  default:
+    - udp:8.8.8.8
+    - wrongprotocol:8.8.4.4
+    - udp:1.1.1.1`
+				helpertest.ShouldLogFatal(func() {
+					unmarshalConfig([]byte(data), cfg)
+				})
+			})
+		})
+
+		When("config is not YAML", func() {
+			It("should log with fatal and exit", func() {
+				cfg := Config{}
+				data :=
+					`///`
+				helpertest.ShouldLogFatal(func() {
+					unmarshalConfig([]byte(data), cfg)
+				})
+			})
+		})
+
+		When("deprecated querylog.dir parameter is used", func() {
+			It("should be mapped to csv writer", func() {
+				By("per client", func() {
+					c := &Config{
+						QueryLog: QueryLogConfig{
+							Dir:       "/somedir",
+							PerClient: true,
+						}}
+					validateConfig(c)
+
+					Expect(c.QueryLog.Target).Should(Equal("/somedir"))
+					Expect(c.QueryLog.Type).Should(Equal(QueryLogTypeCsvClient))
+				})
+
+				By("one file", func() {
+					c := &Config{
+						QueryLog: QueryLogConfig{
+							Dir:       "/somedir",
+							PerClient: false,
+						}}
+					validateConfig(c)
+
+					Expect(c.QueryLog.Target).Should(Equal("/somedir"))
+					Expect(c.QueryLog.Type).Should(Equal(QueryLogTypeCsv))
+				})
+
+			})
+		})
+
 		When("config directory does not exist", func() {
 			It("should log with fatal and exit if config is mandatory", func() {
 				err := os.Chdir("../..")
@@ -75,7 +168,7 @@ var _ = Describe("Config", func() {
 				var fatal bool
 
 				Log().ExitFunc = func(int) { fatal = true }
-				_ = NewConfig("config.yml", true)
+				LoadConfig("config.yml", true)
 
 				Expect(fatal).Should(BeTrue())
 			})
@@ -84,9 +177,9 @@ var _ = Describe("Config", func() {
 				err := os.Chdir("../..")
 				Expect(err).Should(Succeed())
 
-				cfg := NewConfig("config.yml", false)
+				LoadConfig("config.yml", false)
 
-				Expect(cfg.LogLevel).Should(Equal("info"))
+				Expect(config.LogLevel).Should(Equal(LevelInfo))
 			})
 		})
 	})
@@ -103,63 +196,63 @@ var _ = Describe("Config", func() {
 		},
 		Entry("udp with port",
 			"udp:4.4.4.4:531",
-			Upstream{Net: "tcp+udp", Host: "4.4.4.4", Port: 531},
+			Upstream{Net: NetProtocolTcpUdp, Host: "4.4.4.4", Port: 531},
 			false),
 		Entry("udp without port, use default",
 			"udp:4.4.4.4",
-			Upstream{Net: "tcp+udp", Host: "4.4.4.4", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "4.4.4.4", Port: 53},
 			false),
 		Entry("tcp with port",
 			"tcp:4.4.4.4:4711",
-			Upstream{Net: "tcp+udp", Host: "4.4.4.4", Port: 4711},
+			Upstream{Net: NetProtocolTcpUdp, Host: "4.4.4.4", Port: 4711},
 			false),
 		Entry("tcp without port, use default",
 			"tcp:4.4.4.4",
-			Upstream{Net: "tcp+udp", Host: "4.4.4.4", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "4.4.4.4", Port: 53},
 			false),
 		Entry("tcp-tls without port, use default",
 			"tcp-tls:4.4.4.4",
-			Upstream{Net: "tcp-tls", Host: "4.4.4.4", Port: 853},
+			Upstream{Net: NetProtocolTcpTls, Host: "4.4.4.4", Port: 853},
 			false),
 		Entry("DoH without port, use default",
 			"https:4.4.4.4",
-			Upstream{Net: "https", Host: "4.4.4.4", Port: 443},
+			Upstream{Net: NetProtocolHttps, Host: "4.4.4.4", Port: 443},
 			false),
 		Entry("DoH with port",
 			"https:4.4.4.4:888",
-			Upstream{Net: "https", Host: "4.4.4.4", Port: 888},
+			Upstream{Net: NetProtocolHttps, Host: "4.4.4.4", Port: 888},
 			false),
 		Entry("DoH named",
 			"https://dns.google/dns-query",
-			Upstream{Net: "https", Host: "dns.google", Port: 443, Path: "/dns-query"},
+			Upstream{Net: NetProtocolHttps, Host: "dns.google", Port: 443, Path: "/dns-query"},
 			false),
 		Entry("DoH named, path with multiple slashes",
 			"https://dns.google/dns-query/a/b",
-			Upstream{Net: "https", Host: "dns.google", Port: 443, Path: "/dns-query/a/b"},
+			Upstream{Net: NetProtocolHttps, Host: "dns.google", Port: 443, Path: "/dns-query/a/b"},
 			false),
 		Entry("DoH named with port",
 			"https://dns.google:888/dns-query",
-			Upstream{Net: "https", Host: "dns.google", Port: 888, Path: "/dns-query"},
+			Upstream{Net: NetProtocolHttps, Host: "dns.google", Port: 888, Path: "/dns-query"},
 			false),
 		Entry("empty",
 			"",
-			Upstream{Net: ""},
+			Upstream{Net: 0},
 			false),
 		Entry("udpIpv6WithPort",
 			"udp:[fd00::6cd4:d7e0:d99d:2952]:53",
-			Upstream{Net: "tcp+udp", Host: "fd00::6cd4:d7e0:d99d:2952", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "fd00::6cd4:d7e0:d99d:2952", Port: 53},
 			false),
 		Entry("udpIpv6WithPort2",
 			"udp://[2001:4860:4860::8888]:53",
-			Upstream{Net: "tcp+udp", Host: "2001:4860:4860::8888", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "2001:4860:4860::8888", Port: 53},
 			false),
 		Entry("default net, default port",
 			"1.1.1.1",
-			Upstream{Net: "tcp+udp", Host: "1.1.1.1", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "1.1.1.1", Port: 53},
 			false),
 		Entry("default net with port",
 			"1.1.1.1:153",
-			Upstream{Net: "tcp+udp", Host: "1.1.1.1", Port: 153},
+			Upstream{Net: NetProtocolTcpUdp, Host: "1.1.1.1", Port: 153},
 			false),
 		Entry("with negative port",
 			"tcp:4.4.4.4:-1",
@@ -179,11 +272,11 @@ var _ = Describe("Config", func() {
 			true),
 		Entry("tcp+udp",
 			"tcp+udp:1.1.1.1:53",
-			Upstream{Net: "tcp+udp", Host: "1.1.1.1", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "1.1.1.1", Port: 53},
 			false),
 		Entry("tcp+udp default port",
 			"tcp+udp:1.1.1.1",
-			Upstream{Net: "tcp+udp", Host: "1.1.1.1", Port: 53},
+			Upstream{Net: NetProtocolTcpUdp, Host: "1.1.1.1", Port: 53},
 			false),
 	)
 })
