@@ -1,8 +1,7 @@
 package lists
 
+//go:generate go-enum -f=$GOFILE --marshal --names
 import (
-	"blocky/evt"
-	"blocky/util"
 	"bufio"
 	"errors"
 	"fmt"
@@ -10,71 +9,20 @@ import (
 	"net"
 	"net/http"
 	"os"
-	"sort"
 	"strings"
 	"sync"
 	"time"
 
-	"blocky/log"
-
+	"github.com/0xERR0R/blocky/evt"
+	"github.com/0xERR0R/blocky/log"
 	"github.com/sirupsen/logrus"
 )
 
-const (
-	defaultRefreshPeriod = 4 * time.Hour
-)
-
-// ListCacheType represents the type of a cached list
+// ListCacheType represents the type of cached list ENUM(
+// blacklist // is a list with blocked domains
+// whitelist // is a list with whitelisted domains / IPs
+// )
 type ListCacheType int
-
-const (
-	// BLACKLIST is a list with blocked domains / IPs
-	BLACKLIST ListCacheType = iota
-
-	// WHITELIST is a list with whitelisted domains / IPs
-	WHITELIST
-)
-
-func (l ListCacheType) String() string {
-	names := [...]string{
-		"blacklist",
-		"whitelist"}
-
-	return names[l]
-}
-
-// nolint:gochecknoglobals
-var timeout = 60 * time.Second
-
-type stringCache map[int]string
-
-func (cache stringCache) elementCount() int {
-	count := 0
-
-	for k, v := range cache {
-		count += len(v) / k
-	}
-
-	return count
-}
-
-func (cache stringCache) contains(searchString string) bool {
-	searchLen := len(searchString)
-	if searchLen == 0 {
-		return false
-	}
-
-	searchBucketLen := len(cache[searchLen]) / searchLen
-	idx := sort.Search(searchBucketLen, func(i int) bool {
-		return cache[searchLen][i*searchLen:i*searchLen+searchLen] >= searchString
-	})
-
-	if idx < searchBucketLen {
-		return cache[searchLen][idx*searchLen:idx*searchLen+searchLen] == strings.ToLower(searchString)
-	}
-
-	return false
-}
 
 // Matcher checks if a domain is in a list
 type Matcher interface {
@@ -87,18 +35,19 @@ type Matcher interface {
 
 // ListCache generic cache of strings divided in groups
 type ListCache struct {
-	groupCaches map[string]stringCache
+	groupCaches map[string]cache
 	lock        sync.RWMutex
 
-	groupToLinks  map[string][]string
-	refreshPeriod time.Duration
-	listType      ListCacheType
+	groupToLinks    map[string][]string
+	refreshPeriod   time.Duration
+	downloadTimeout time.Duration
+	listType        ListCacheType
 }
 
 // Configuration returns current configuration and stats
 func (b *ListCache) Configuration() (result []string) {
 	if b.refreshPeriod > 0 {
-		result = append(result, fmt.Sprintf("refresh period: %d minutes", b.refreshPeriod/time.Minute))
+		result = append(result, fmt.Sprintf("refresh period: %s", b.refreshPeriod))
 	} else {
 		result = append(result, "refresh: disabled")
 	}
@@ -106,7 +55,12 @@ func (b *ListCache) Configuration() (result []string) {
 	result = append(result, "group links:")
 	for group, links := range b.groupToLinks {
 		result = append(result, fmt.Sprintf("  %s:", group))
+
 		for _, link := range links {
+			if strings.Contains(link, "\n") {
+				link = "[INLINE DEFINITION]"
+			}
+
 			result = append(result, fmt.Sprintf("   - %s", link))
 		}
 	}
@@ -122,23 +76,20 @@ func (b *ListCache) Configuration() (result []string) {
 
 	result = append(result, fmt.Sprintf("  TOTAL: %d entries", total))
 
-	return
+	return result
 }
 
 // NewListCache creates new list instance
-func NewListCache(t ListCacheType, groupToLinks map[string][]string, refreshPeriod int) *ListCache {
-	groupCaches := make(map[string]stringCache)
-
-	p := time.Duration(refreshPeriod) * time.Minute
-	if refreshPeriod == 0 {
-		p = defaultRefreshPeriod
-	}
+func NewListCache(t ListCacheType, groupToLinks map[string][]string, refreshPeriod time.Duration,
+	downloadTimeout time.Duration) *ListCache {
+	groupCaches := make(map[string]cache)
 
 	b := &ListCache{
-		groupToLinks:  groupToLinks,
-		groupCaches:   groupCaches,
-		refreshPeriod: p,
-		listType:      t,
+		groupToLinks:    groupToLinks,
+		groupCaches:     groupCaches,
+		refreshPeriod:   refreshPeriod,
+		downloadTimeout: downloadTimeout,
+		listType:        t,
 	}
 	b.Refresh()
 
@@ -165,24 +116,21 @@ func logger() *logrus.Entry {
 }
 
 // downloads and reads files with domain names and creates cache for them
-func createCacheForGroup(links []string) stringCache {
-	cache := make(stringCache)
-
-	keys := make(map[string]struct{})
-
+func (b *ListCache) createCacheForGroup(links []string) cache {
 	var wg sync.WaitGroup
 
 	c := make(chan []string, len(links))
 
+	// loop over links (http/local) or inline definitions
 	for _, link := range links {
 		wg.Add(1)
 
-		go processFile(link, c, &wg)
+		go b.processFile(link, c, &wg)
 	}
 
 	wg.Wait()
 
-	tmp := make(map[int]*strings.Builder)
+	factory := newChainedCacheFactory()
 
 Loop:
 	for {
@@ -192,13 +140,7 @@ Loop:
 				return nil
 			}
 			for _, entry := range res {
-				if _, value := keys[entry]; !value {
-					keys[entry] = struct{}{}
-					if tmp[len(entry)] == nil {
-						tmp[len(entry)] = &strings.Builder{}
-					}
-					tmp[len(entry)].WriteString(entry)
-				}
+				factory.addEntry(entry)
 			}
 		default:
 			close(c)
@@ -206,16 +148,7 @@ Loop:
 		}
 	}
 
-	for k, v := range tmp {
-		chunks := util.Chunks(v.String(), k)
-		sort.Strings(chunks)
-
-		cache[k] = strings.Join(chunks, "")
-
-		v.Reset()
-	}
-
-	return cache
+	return factory.create()
 }
 
 // Match matches passed domain name against cached list entries
@@ -224,7 +157,7 @@ func (b *ListCache) Match(domain string, groupsToCheck []string) (found bool, gr
 	defer b.lock.RUnlock()
 
 	for _, g := range groupsToCheck {
-		if b.groupCaches[g].contains(domain) {
+		if c, ok := b.groupCaches[g]; ok && c.contains(domain) {
 			return true, g
 		}
 	}
@@ -235,7 +168,7 @@ func (b *ListCache) Match(domain string, groupsToCheck []string) (found bool, gr
 // Refresh triggers the refresh of a list
 func (b *ListCache) Refresh() {
 	for group, links := range b.groupToLinks {
-		cacheForGroup := createCacheForGroup(links)
+		cacheForGroup := b.createCacheForGroup(links)
 
 		if cacheForGroup != nil {
 			b.lock.Lock()
@@ -254,9 +187,9 @@ func (b *ListCache) Refresh() {
 	}
 }
 
-func downloadFile(link string) (io.ReadCloser, error) {
+func (b *ListCache) downloadFile(link string) (io.ReadCloser, error) {
 	client := http.Client{
-		Timeout: timeout,
+		Timeout: b.downloadTimeout,
 	}
 
 	var resp *http.Response
@@ -301,7 +234,7 @@ func readFile(file string) (io.ReadCloser, error) {
 }
 
 // downloads file (or reads local file) and writes file content as string array in the channel
-func processFile(link string, ch chan<- []string, wg *sync.WaitGroup) {
+func (b *ListCache) processFile(link string, ch chan<- []string, wg *sync.WaitGroup) {
 	defer wg.Done()
 
 	result := make([]string, 0)
@@ -310,9 +243,15 @@ func processFile(link string, ch chan<- []string, wg *sync.WaitGroup) {
 
 	var err error
 
-	if strings.HasPrefix(link, "http") {
-		r, err = downloadFile(link)
-	} else {
+	switch {
+	// link contains a line break -> this is inline list definition in YAML (with literal style Block Scalar)
+	case strings.ContainsAny(link, "\n"):
+		r = io.NopCloser(strings.NewReader(link))
+	// link is http(s) -> download it
+	case strings.HasPrefix(link, "http"):
+		r, err = b.downloadFile(link)
+	// probably path to a local file
+	default:
 		r, err = readFile(link)
 	}
 
@@ -336,7 +275,7 @@ func processFile(link string, ch chan<- []string, wg *sync.WaitGroup) {
 	scanner := bufio.NewScanner(r)
 
 	for scanner.Scan() {
-		line := scanner.Text()
+		line := strings.TrimSpace(scanner.Text())
 		// skip comments
 		if line := processLine(line); line != "" {
 			result = append(result, line)
@@ -372,7 +311,7 @@ func processLine(line string) string {
 			return ip.String()
 		}
 
-		return strings.ToLower(host)
+		return strings.TrimSpace(strings.ToLower(host))
 	}
 
 	return ""
