@@ -1,10 +1,11 @@
 package resolver
 
 import (
-	"blocky/config"
-	. "blocky/helpertest"
 	"net"
 
+	"github.com/0xERR0R/blocky/config"
+	. "github.com/0xERR0R/blocky/helpertest"
+	. "github.com/0xERR0R/blocky/model"
 	"github.com/miekg/dns"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
