@@ -1,14 +1,15 @@
 package resolver
 
 import (
-	"blocky/config"
-	. "blocky/helpertest"
-	"blocky/util"
 	"crypto/tls"
 	"fmt"
 	"net/http"
 	"time"
 
+	"github.com/0xERR0R/blocky/config"
+	. "github.com/0xERR0R/blocky/helpertest"
+	. "github.com/0xERR0R/blocky/model"
+	"github.com/0xERR0R/blocky/util"
 	"github.com/miekg/dns"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -31,7 +32,7 @@ var _ = Describe("UpstreamResolver", func() {
 				resp, err := sut.Resolve(newRequest("example.com.", dns.TypeA))
 				Expect(err).Should(Succeed())
 				Expect(resp.Res.Rcode).Should(Equal(dns.RcodeSuccess))
-				Expect(resp.RType).Should(Equal(RESOLVED))
+				Expect(resp.RType).Should(Equal(ResponseTypeRESOLVED))
 				Expect(resp.Res.Answer).Should(BeDNSRecord("example.com.", dns.TypeA, 123, "123.124.122.122"))
 				Expect(resp.Reason).Should(Equal(fmt.Sprintf("RESOLVED (%s:%d)", upstream.Host, upstream.Port)))
 			})
@@ -49,7 +50,7 @@ var _ = Describe("UpstreamResolver", func() {
 				resp, err := sut.Resolve(newRequest("example.com.", dns.TypeA))
 				Expect(err).Should(Succeed())
 				Expect(resp.Res.Rcode).Should(Equal(dns.RcodeNameError))
-				Expect(resp.RType).Should(Equal(RESOLVED))
+				Expect(resp.RType).Should(Equal(ResponseTypeRESOLVED))
 				Expect(resp.Reason).Should(Equal(fmt.Sprintf("RESOLVED (%s:%d)", upstream.Host, upstream.Port)))
 			})
 		})
@@ -91,7 +92,7 @@ var _ = Describe("UpstreamResolver", func() {
 					Expect(err).Should(Succeed())
 					Expect(resp.Res.Rcode).Should(Equal(dns.RcodeSuccess))
 					Expect(resp.Res.Answer).Should(BeDNSRecord("example.com.", dns.TypeA, 123, "123.124.122.122"))
-					Expect(resp.RType).Should(Equal(RESOLVED))
+					Expect(resp.RType).Should(Equal(ResponseTypeRESOLVED))
 				})
 
 				By("3 attempts with timeout -> should return error", func() {
@@ -140,7 +141,7 @@ var _ = Describe("UpstreamResolver", func() {
 				resp, err := sut.Resolve(newRequest("example.com.", dns.TypeA))
 				Expect(err).Should(Succeed())
 				Expect(resp.Res.Rcode).Should(Equal(dns.RcodeSuccess))
-				Expect(resp.RType).Should(Equal(RESOLVED))
+				Expect(resp.RType).Should(Equal(ResponseTypeRESOLVED))
 				Expect(resp.Res.Answer).Should(BeDNSRecord("example.com.", dns.TypeA, 123, "123.124.122.122"))
 				Expect(resp.Reason).Should(Equal(fmt.Sprintf("RESOLVED (https://%s:%d)", upstream.Host, upstream.Port)))
 			})
@@ -183,7 +184,7 @@ var _ = Describe("UpstreamResolver", func() {
 		})
 		When("Configured DOH resolver does not respond", func() {
 			JustBeforeEach(func() {
-				sut = NewUpstreamResolver(config.Upstream{Net: "https", Host: "wronghost.example.com"})
+				sut = NewUpstreamResolver(config.Upstream{Net: config.NetProtocolHttps, Host: "wronghost.example.com"})
 			})
 			It("should return error", func() {
 				_, err := sut.Resolve(newRequest("example.com.", dns.TypeA))
@@ -193,7 +194,7 @@ var _ = Describe("UpstreamResolver", func() {
 		})
 		When("Configured DOH resolver receives wrong request", func() {
 			JustBeforeEach(func() {
-				sut = NewUpstreamResolver(config.Upstream{Net: "https", Host: "host"})
+				sut = NewUpstreamResolver(config.Upstream{Net: config.NetProtocolHttps, Host: "host"})
 			})
 			It("should return error", func() {
 				wrongReq := new(dns.Msg)
